refactor: Move primary panels into subpackage
