+++ src/main/java/me/coley/recaf/ui/controls/pane/ConfigTabs.java
@@ -1,4 +1,4 @@
-package me.coley.recaf.ui.controls;
+package me.coley.recaf.ui.controls.pane;
 
 import javafx.scene.control.Tab;
 import javafx.scene.control.TabPane;
