+++ src/main/java/me/coley/recaf/ui/controls/pane/ContactInfoPane.java
@@ -1,4 +1,4 @@
-package me.coley.recaf.ui.controls;
+package me.coley.recaf.ui.controls.pane;
 
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
@@ -6,6 +6,7 @@ import javafx.scene.Node;
 import javafx.scene.control.Hyperlink;
 import javafx.scene.control.Label;
 import javafx.scene.layout.GridPane;
+import me.coley.recaf.ui.controls.IconView;
 import me.coley.recaf.util.Log;
 import java.awt.*;
 import java.io.IOException;
